Create PanelSection and adjust UI elements
