+++ src/views/FoldersView.tsx
@@ -1,11 +1,9 @@
 import { FileEntry } from '@tauri-apps/api/fs';
 import { useEffect, useState } from 'react';
-import { FileUploader } from 'react-drag-drop-files';
 
+import { PanelSection } from '../Components/PanelSection';
 import { cx } from '../cx';
-import { ensureProjectFileStructure, readAllProjectFiles, runPDFIngestion, uploadFiles } from '../filesystem';
-
-const BASE_DIR = await ensureProjectFileStructure();
+import { readAllProjectFiles } from '../filesystem';
 
 export function FoldersView({ onClick }: { onClick?: (fileEntry: FileEntry) => void }) {
   const [files, setFiles] = useState<FileEntry[]>([]);
@@ -31,32 +29,15 @@ export function FoldersView({ onClick }: { onClick?: (fileEntry: FileEntry) => v
     onClick?.(file);
   }
 
-  async function handleChange(uploadedFiles: FileList) {
-    try {
-      await uploadFiles(uploadedFiles);
-      console.log('File uploaded with success');
-
-      const projectFiles = await readAllProjectFiles();
-      setFiles(projectFiles);
-
-      await runPDFIngestion();
-      console.log('PDFs ingested with success');
-    } catch (err) {
-      console.error('Error uploading references', err);
-    }
-  }
-
   return (
-    <div className="flex h-full flex-col justify-between px-2 pt-2">
-      <div className="flex h-full w-full flex-1 flex-col">
-        <div className="mb-2 flex flex-col text-sm font-bold uppercase">Project X</div>
+    <>
+      <PanelSection title="Open Files">
+        <div />
+      </PanelSection>
+      <PanelSection title="Project X">
         <FileTree files={files} root selectedFile={selectedFile} onClick={handleOnClick} />
-      </div>
-      <div>
-        <FileUploader handleChange={handleChange} label="Upload or drop a file right here" multiple name="file" />
-        <code className="mb-auto block text-xs font-normal">{BASE_DIR}</code>
-      </div>
-    </div>
+      </PanelSection>
+    </>
   );
 }
 
@@ -69,7 +50,7 @@ interface FileTreeBaseProps {
 }
 
 const FileTree = ({ files, root, ...props }: FileTreeBaseProps) => (
-  <div className="flex-1 overflow-scroll">
+  <div className="xflex-1 overflow-scroll">
     <ul
       className={cx('', {
         'ml-6': !root,
